+++ tests/testthat/test-6-plans.R
@@ -802,3 +802,7 @@ test_with_dir("trailing dots in imports (#1147)", {
 test_with_dir("type_sum() S3 method for printing language columns", {
   expect_equal(type_sum.expr_list("123"), "expr")
 })
+
+test_with_dir("illegal plan error message (#1334)", {
+  expect_error(make(plan = function(x) x), regexp = "drake plan")
+})